separate qiita api function
