+++ front/src/App.tsx
@@ -2,7 +2,7 @@ import Activity from "./components/Activity";
 import Certification from "./components/Certification";
 import Github from "./components/github/Github";
 import Profile from "./components/Profile";
-import QiitaArt from "./components/Qiita";
+import Qiita from "./components/qiita/Qiita";
 import Skill from "./components/SkillSet";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
@@ -16,7 +16,7 @@ function App() {
       <Certification />
       <Activity />
       <Github />
-      <QiitaArt />
+      <Qiita />
       <Footer />
     </div>
   )
